[FIX] coding

+++ config.h
@@ -1,5 +1,7 @@
 /* See LICENSE file for copyright and license details. */
 
+#include <stdint.h>
+
 
 #define MY_STATUS \
     sprintf(stext, "dwm-%s",   \
@@ -16,8 +18,8 @@ static const char normfgcolor[]     = "#bbbbbb";
 static const char selbordercolor[]  = "#005577";
 static const char selbgcolor[]      = "#005577";
 static const char selfgcolor[]      = "#eeeeee";
-static const unsigned int borderpx  = 1;        /* border pixel of windows */
-static const unsigned int snap      = 32;       /* snap pixel */
+static const uint32_t borderpx  = 1;        /* border pixel of windows */
+static const uint32_t snap      = 32;       /* snap pixel */
 static const int showbar            = 1;        /* 0 means no bar */
 static const int topbar             = 1;        /* 0 means bottom bar */
 
@@ -45,7 +47,7 @@ static const Rule rules[] = {
 
 /* layout(s) */
 static const float mfact     = 0.55; /* factor of master area size [0.05..0.95] */
-static const int nmaster     = 1;    /* number of clients in master area */
+static const size_t nmaster     = 1;    /* number of clients in master area */
 static const int resizehints = False;    /* 1 means respect size hints in tiled resizals */
 
 static const Layout layouts[] = {
